ConditionParser refactor and fix "this OR" conditions
